+++ storm-core/src/jvm/org/apache/storm/testing/NGrouping.java
@@ -28,9 +28,7 @@ public class NGrouping implements CustomStreamGrouping {
     int _n;
     List<Integer> _outTasks;
     
-    public NGrouping(int n) {
-        _n = n;
-    }
+    public NGrouping(Integer n) {_n = n;}
     
     @Override
     public void prepare(WorkerTopologyContext context, GlobalStreamId stream, List<Integer> targetTasks) {